finish all

+++ src/main/java/com/collegegroup/processscheduling/Comparators/SortByFCFS.java
@@ -10,6 +10,6 @@ public class SortByFCFS implements Comparator<GUIProcess> {
     // roll number
     public int compare(GUIProcess a, GUIProcess b)
     {
-        return (int) (Double.parseDouble(a.getArrivalTime()) - Double.parseDouble(b.getArrivalTime()));
+        return Integer.parseInt(a.getArrivalTime()) - Integer.parseInt(b.getArrivalTime());
     }
 }
